Extract picker loading from ComposerVC into an injectable loader
